Refactor gaussian subpackage and get spherical overlaps working
