+++ moldesign/orbitals/__init__.py
@@ -4,6 +4,8 @@ def toplevel(o):
 __all__ = []
 
 from .gaussians import *
+from .cartesian import *
+from .spherical import *
 from .orbitals import *
 from .atomic_basis_fn import *
 from .basis import *
